% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generalUtils.R
\docType{data}
\name{appConfig}
\alias{appConfig}
\title{load app configuration (schedules, thresholds)}
\format{An object of class \code{data.frame} with 18 rows and 2 columns.}
\usage{
appConfig
}
\description{
...
}
\keyword{datasets}
