% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generalUtils.R
\name{compress}
\alias{compress}
\title{This function takes a number and returns a compressed string (e.g. 1624 => 1.6K or 2K, depending on round.by)}
\usage{
compress(x, round.by = 2)
}
\description{
...
}
