% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_generalUtils.R
\docType{data}
\name{dbConfig}
\alias{dbConfig}
\title{reactive object for database credentials}
\format{An object of class \code{list} of length 2.}
\usage{
dbConfig
}
\description{
...
}
\keyword{datasets}

