% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qry95thPercStmt}
\alias{qry95thPercStmt}
\title{Statements with Runtime in 95th Percentile}
\usage{
qry95thPercStmt()
}
\description{
Function to query the slowest statements. The number of results can be configured through
a limit in the configuration file - cnf.file - (qryFullTableScanLimit).
}
