% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryErrWarnData}
\alias{qryErrWarnData}
\title{Statements with warnings or errors.}
\usage{
qryErrWarnData()
}
\description{
Function to query statements with errors or warnings. The number of results can be configured through a limit in the
configuration file - cnf.file - (qryFilterErrors, qryFilterWarnings). Statements with more errors or warnings then the threshold
will be in the result set.
}
