% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryFlagTokuEngine}
\alias{qryFlagTokuEngine}
\title{Check if TokuDB-Storage-Engine is in use}
\usage{
qryFlagTokuEngine()
}
\description{
Function to check if TokuDB-Storage-Engine is in use
}
