% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryFullTblScanStmt}
\alias{qryFullTblScanStmt}
\title{Statements with full table scan}
\usage{
qryFullTblScanStmt()
}
\description{
Function to query statements with full table scan. The number of results is limited to 50.
}

