% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryStmtAnalysis}
\alias{qryStmtAnalysis}
\title{Statement Analysis}
\usage{
qryStmtAnalysis()
}
\description{
Function to query the most often used statements. The number of results can be configured through a
limit in the configuration file - cnf.file - (qryStatementAnalysisLimit).
}
