% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryTimeLine}
\alias{qryTimeLine}
\title{Aggregated processlist}
\usage{
qryTimeLine()
}
\description{
Function to query semi-aggregated process data of mariadb. Output of the function is an xts object.
}

