% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_mariaDBqueryFunctions.R
\name{qryUserStat}
\alias{qryUserStat}
\title{User Statistic}
\usage{
qryUserStat()
}
\description{
Function to query statistics of the mariadb users. The number of results can be configured through
a limit in the configuration file - cnf.file - (qryUserStatLimit).
}

