% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAMESPACE.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{renderTable}
\alias{reexports}
\alias{box}
\alias{reexports}
\alias{dashboardHeader}
\alias{reexports}
\alias{notificationItem}
\alias{reexports}
\alias{dashboardSidebar}
\alias{reexports}
\alias{dashboardBody}
\alias{reexports}
\alias{dashboardPage}
\alias{reexports}
\alias{dropdownMenuOutput}
\alias{reexports}
\alias{dropdownMenu}
\alias{reexports}
\alias{sidebarSearchForm}
\alias{reexports}
\alias{tabItem}
\alias{reexports}
\alias{tabBox}
\alias{reexports}
\alias{sidebarMenu}
\alias{reexports}
\alias{menuItem}
\alias{reexports}
\alias{tabItems}
\alias{reexports}
\alias{gvisBarChart}
\alias{reexports}
\alias{renderGvis}
\alias{reexports}
\alias{gvisPieChart}
\alias{reexports}
\alias{formatIECBytes}
\alias{reexports}
\alias{seconds_to_period}
\alias{reexports}
\alias{dygraphOutput}
\alias{reexports}
\alias{renderDygraph}
\alias{reexports}
\alias{dygraph}
\alias{reexports}
\alias{dyAxis}
\alias{reexports}
\alias{dySeries}
\alias{reexports}
\alias{dyOptions}
\alias{reexports}
\alias{dyLegend}
\alias{reexports}
\alias{dyRangeSelector}
\alias{reexports}
\alias{dyHighlight}
\alias{reexports}
\alias{renderRHandsontable}
\alias{reexports}
\alias{rhandsontable}
\alias{reexports}
\alias{hot_col}
\alias{reexports}
\alias{rHandsontableOutput}
\alias{reexports}
\alias{hot_to_r}
\alias{reexports}
\alias{xts}
\alias{reexports}
\alias{datatable}
\alias{reexports}
\alias{formatCurrency}
\alias{reexports}
\alias{formatStyle}
\alias{reexports}
\alias{dataTableOutput}
\alias{reexports}
\alias{styleColorBar}
\alias{reexports}
\alias{styleInterval}
\alias{reexports}
\alias{formatRound}
\alias{reexports}
\alias{renderDataTable}
\alias{reexports}
\alias{dataTableProxy}
\alias{reexports}
\alias{replaceData}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{spread}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}}

  \item{DT}{\code{\link[DT]{datatable}}, \code{\link[DT]{formatCurrency}}, \code{\link[DT]{formatStyle}}, \code{\link[DT]{dataTableOutput}}, \code{\link[DT]{styleColorBar}}, \code{\link[DT]{styleInterval}}, \code{\link[DT]{formatRound}}, \code{\link[DT]{renderDataTable}}, \code{\link[DT]{dataTableProxy}}, \code{\link[DT]{replaceData}}}

  \item{dygraphs}{\code{\link[dygraphs]{dygraphOutput}}, \code{\link[dygraphs]{renderDygraph}}, \code{\link[dygraphs]{dygraph}}, \code{\link[dygraphs]{dyAxis}}, \code{\link[dygraphs]{dySeries}}, \code{\link[dygraphs]{dyOptions}}, \code{\link[dygraphs]{dyLegend}}, \code{\link[dygraphs]{dyRangeSelector}}, \code{\link[dygraphs]{dyHighlight}}}

  \item{googleVis}{\code{\link[googleVis]{gvisBarChart}}, \code{\link[googleVis]{renderGvis}}, \code{\link[googleVis]{gvisPieChart}}}

  \item{humanFormat}{\code{\link[humanFormat]{formatIECBytes}}}

  \item{lubridate}{\code{\link[lubridate]{seconds_to_period}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rhandsontable}{\code{\link[rhandsontable]{renderRHandsontable}}, \code{\link[rhandsontable]{rhandsontable}}, \code{\link[rhandsontable]{hot_col}}, \code{\link[rhandsontable]{rHandsontableOutput}}, \code{\link[rhandsontable]{hot_to_r}}}

  \item{shiny}{\code{\link[shiny]{renderTable}}}

  \item{shinydashboard}{\code{\link[shinydashboard]{box}}, \code{\link[shinydashboard]{dashboardHeader}}, \code{\link[shinydashboard]{notificationItem}}, \code{\link[shinydashboard]{dashboardSidebar}}, \code{\link[shinydashboard]{dashboardBody}}, \code{\link[shinydashboard]{dashboardPage}}, \code{\link[shinydashboard]{dropdownMenuOutput}}, \code{\link[shinydashboard]{dropdownMenu}}, \code{\link[shinydashboard]{sidebarSearchForm}}, \code{\link[shinydashboard]{tabItem}}, \code{\link[shinydashboard]{tabBox}}, \code{\link[shinydashboard]{sidebarMenu}}, \code{\link[shinydashboard]{menuItem}}, \code{\link[shinydashboard]{tabItems}}}

  \item{tidyr}{\code{\link[tidyr]{spread}}}

  \item{xts}{\code{\link[xts]{xts}}}
}}

